refactor: 좋아요한 대학지원정보 목록 조회 함수 이동
- MyPageService에서 UnivApplyInfoLikeService 로
